CU-86ennrjnk Created new pages and linked with routings
