+++ src/views/findCourse/findCourse.tsx
@@ -1,0 +1,5 @@
+const findCourse = () => {
+  return <div>findCourse</div>;
+};
+
+export default findCourse;
