+++ src/views/index.tsx
@@ -1,0 +1,3 @@
+export { default as Home } from "./home/Home";
+export { default as FindCourses } from "./findCourse/findCourse";
+export { default as Institutes } from "./institutes/Institutes";
