+++ src/views/institutes/Institutes.tsx
@@ -1,0 +1,5 @@
+const Institutes = () => {
+  return <div>Institutes</div>;
+};
+
+export default Institutes;
